requirements.txt
python-dotenv>=1.0.1
requests>=2.31.0
urllib3>=2.2.2
python-dateutil>=2.9.0.post0
pydexcom>=0.3.2
pandas>=2.2.2
matplotlib>=3.9.0
